requirements.txt
numpy==1.21.2
Pillow==8.3.2
tqdm==4.62.2
